neuroscribe-pdf-copilot/requirements.txt
streamlit
PyMuPDF
Pillow
python-dotenv
openai